Modified all credit amounts to decimal type Added some extra checks for resume game session and play live #35

+++ resources/assets/js/pages/manage_session.js
@@ -9,7 +9,7 @@ $(document).ready(function() {
             coins = 0;
         }
 
-        let amount = parseInt(coins) * rate;
+        let amount = parseFloat(coins) * rate;
 
         $inputMoney.show();
         $inputMoney.find('span').html(amount.toFixed(2));
